+++ usrsctplib/netinet/sctp_bsd_addr.c
@@ -800,7 +800,7 @@ sctp_add_or_del_interfaces(int (*pred)(struct ifnet *), int add)
 #endif
 
 struct mbuf *
-sctp_get_mbuf_for_msg(unsigned int space_needed, int want_header,
+sctp_get_mbuf_for_msg(size_t space_needed, int want_header,
 		      int how, int allonebuf, int type)
 {
     struct mbuf *m = NULL;
